Rename "avocado.core.status" to "avocado.core.teststatus"
And free the "status" namespace to a more complex module structure to
handle the N(ext) Runner status utilities.
